[refactor] : 예약 목록 페이지 구조 정비 및 UI 컴포넌트 분리
- ReservationListView 내 반복 UI 분리: ReservationCard 컴포넌트 생성
- 상태별 버튼 영역 분리: ReservationActionButtons 생성 (status 조건 분기)
- 예약 DTO 타입을 ReservationDTO, ReservationCardProps로 분리
- status 리터럴 타입 불일치 문제 해결: mapReservationToCardProps 유틸 추가
- 불필요한 status string → 유효한 status union 타입으로 좁힘
- useReservatedList 훅 내부 상태를 View에 props로만 전달되도록 리팩토링
- UI와 로직 책임 분리: View는 순수 렌더링만 담당하도록 구조 개선
