+++ src/types/Reservation.type.ts
@@ -1,8 +1,11 @@
-export interface ReservationData {
-  studioId: string;
-  profileImage: string;
-  reservationId: string;
-  studioname: string;
-  status: boolean;
+export interface ReservationCardProps {
+  reservationId: number;
+  studioId: number;
+  studioName: string;
+  studioImage: string;
+  productName: string;
   createDate: string;
+  createTime: string;
+  status: "예약확정" | "촬영완료" | "예약취소";
+  productId: number;
 }
